Add possibility to use static landing page
